Moved Module and ModuleManager into its own package folder
