+++ src/main/java/seedu/address/logic/commands/ModuleAddCommand.java
@@ -6,8 +6,8 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_MODULE_NAME;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.Module;
-import seedu.address.model.ModuleManager;
+import seedu.address.model.module.Module;
+import seedu.address.model.module.ModuleManager;
 
 /**
  * Adds a module to Trajectory
