+++ src/main/java/seedu/address/logic/commands/ModuleListCommand.java
@@ -2,9 +2,9 @@ package seedu.address.logic.commands;
 
 import seedu.address.logic.CommandHistory;
 import seedu.address.model.Model;
-import seedu.address.model.Module;
-import seedu.address.model.ModuleManager;
 import seedu.address.model.StorageController;
+import seedu.address.model.module.Module;
+import seedu.address.model.module.ModuleManager;
 
 /**
  * Lists all modules in Trajectory to the user.
