+++ src/main/java/seedu/address/model/module/Module.java
@@ -1,4 +1,4 @@
-package seedu.address.model;
+package seedu.address.model.module;
 
 /**
  * Represents a module in Trajectory
