+++ src/main/java/seedu/address/storage/serializable/XmlSerializableModuleList.java
@@ -9,7 +9,7 @@ import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlElementWrapper;
 import javax.xml.bind.annotation.XmlRootElement;
 
-import seedu.address.model.Module;
+import seedu.address.model.module.Module;
 import seedu.address.storage.adapter.XmlAdaptedModule;
 
 /**
